Moved GoToFindController
