+++ app/controllers/candidate_interface/course_choices/go_to_find_controller.rb
@@ -1,0 +1,16 @@
+module CandidateInterface
+  module CourseChoices
+    class GoToFindController < ::CandidateInterface::ContinuousApplicationsController
+      def new
+        @wizard = CourseSelectionWizard.new(current_step:)
+        @adviser_sign_up = Adviser::SignUp.new(current_application)
+      end
+
+    private
+
+      def current_step
+        :go_to_find_explanation
+      end
+    end
+  end
+end
