+++ app/views/candidate_interface/course_choices/go_to_find/new.html.erb
